Add container to arguments passed to `dataGetter`
This PR aligns the `dataGetter` and `cellRenderer` function calls by ensuring the container property is passed.

I would like to be able to access custom props that are passed to the container in the data getter function. Currently this is supported in the `cellRenderer` however we don't need the power of the that function (cell renderering for us is taken care of elsewhere).